app/pyproject.toml
[tool.poetry]
name = "trickstertrust"
version = "0.1.0"
description = ""
authors = ["DarkLorianPrime <lolpokens@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.7"
djangorestframework = "^3.14.0"
redis = "^4.5.4"
celery = "^5.2.7"
asyncpg = "^0.27.0"
python-dotenv = "^1.0.0"
django-extensions = "^3.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
